devkey/demo-2/pyproject.toml
[project]
name = "marathon-agents"
version = "0.1.0"
description = "Multi-agent marathon planning system with ADK, A2A, and Vertex AI Evaluation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform[agent_engines,adk,evaluation]>=1.121.0",
    "google-adk>=1.25.0",
    "a2a-sdk>=0.3.9",
    "pydantic>=2.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "google-auth>=2.0.0",
    "pandas>=2.0.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
default = true

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["src/_agent_template", "node_modules", ".venv"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
